requirements.txt
python-telegram-bot==20.8
google-generativeai==0.4.1
beautifulsoup4==4.12.3
requests==2.31.0
feedparser==6.0.11
python-dotenv==1.0.1
Flask==3.0.2
gunicorn==21.2.0
